api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12